Implemented unsafe insertions, and fixed offset position for inserting values

+++ flitdb/flit.c
@@ -35,7 +35,7 @@ flitdb_extern char *flitdb_errmsg(flitdb **handler)
 	return flitdb_get_err_message(handler);
 }
 
-flitdb_extern int flitdb_extract(flitdb **handler, unsigned short column_position, unsigned short row_position)
+flitdb_extern int flitdb_extract(flitdb **handler, unsigned long long int column_position, unsigned short row_position)
 {
 	return flitdb_read_at(handler, column_position, row_position);
 }
@@ -45,7 +45,7 @@ int flitdb_retrieved_type(flitdb **handler)
 	return flitdb_retrieve_value_type(handler);
 }
 
-flitdb_extern int flitdb_insert_int(flitdb **handler, unsigned short column_position, unsigned short row_position, signed long long int value)
+flitdb_extern int flitdb_insert_int(flitdb **handler, unsigned long long int column_position, unsigned short row_position, signed long long int value)
 {
 	flitdb_insert_value_int(handler, value);
 	return flitdb_insert_at(handler, column_position, row_position);
@@ -53,14 +53,14 @@ flitdb_extern int flitdb_insert_int(flitdb **handler, unsigned short column_posi
 
 #ifdef __cplusplus
 
-int flitdb_insert(flitdb **handler, unsigned short column_position, unsigned short row_position, signed long long int value)
+int flitdb_insert(flitdb **handler, unsigned long long int column_position, unsigned short row_position, signed long long int value)
 {
 	return flitdb_insert_int(handler, column_position, row_position, value);
 }
 
 #endif
 
-flitdb_extern int flitdb_insert_double(flitdb **handler, unsigned short column_position, unsigned short row_position, long double value)
+flitdb_extern int flitdb_insert_double(flitdb **handler, unsigned long long int column_position, unsigned short row_position, long double value)
 {
 	flitdb_insert_value_double(handler, value);
 	return flitdb_insert_at(handler, column_position, row_position);
@@ -68,14 +68,14 @@ flitdb_extern int flitdb_insert_double(flitdb **handler, unsigned short column_p
 
 #ifdef __cplusplus
 
-int flitdb_insert(flitdb **handler, unsigned short column_position, unsigned short row_position, long double value)
+int flitdb_insert(flitdb **handler, unsigned long long int column_position, unsigned short row_position, long double value)
 {
 	return flitdb_insert_double(handler, column_position, row_position, value);
 }
 
 #endif
 
-flitdb_extern int flitdb_insert_float(flitdb **handler, unsigned short column_position, unsigned short row_position, float value)
+flitdb_extern int flitdb_insert_float(flitdb **handler, unsigned long long int column_position, unsigned short row_position, float value)
 {
 	flitdb_insert_value_float(handler, value);
 	return flitdb_insert_at(handler, column_position, row_position);
@@ -83,14 +83,14 @@ flitdb_extern int flitdb_insert_float(flitdb **handler, unsigned short column_po
 
 #ifdef __cplusplus
 
-int flitdb_insert(flitdb **handler, unsigned short column_position, unsigned short row_position, float value)
+int flitdb_insert(flitdb **handler, unsigned long long int column_position, unsigned short row_position, float value)
 {
 	return flitdb_insert_float(handler, column_position, row_position, value);
 }
 
 #endif
 
-flitdb_extern int flitdb_insert_char(flitdb **handler, unsigned short column_position, unsigned short row_position, char *value)
+flitdb_extern int flitdb_insert_char(flitdb **handler, unsigned long long int column_position, unsigned short row_position, char *value)
 {
 	flitdb_insert_value_char(handler, value);
 	return flitdb_insert_at(handler, column_position, row_position);
@@ -98,14 +98,14 @@ flitdb_extern int flitdb_insert_char(flitdb **handler, unsigned short column_pos
 
 #ifdef __cplusplus
 
-int flitdb_insert(flitdb **handler, unsigned short column_position, unsigned short row_position, char *value)
+int flitdb_insert(flitdb **handler, unsigned long long int column_position, unsigned short row_position, char *value)
 {
 	return flitdb_insert_char(handler, column_position, row_position, value);
 }
 
 #endif
 
-flitdb_extern int flitdb_insert_const_char(flitdb **handler, unsigned short column_position, unsigned short row_position, const char *value)
+flitdb_extern int flitdb_insert_const_char(flitdb **handler, unsigned long long int column_position, unsigned short row_position, const char *value)
 {
 	int value_length = strlen(value);
 	char tmp_value[(value_length + 1)];
@@ -117,14 +117,14 @@ flitdb_extern int flitdb_insert_const_char(flitdb **handler, unsigned short colu
 
 #ifdef __cplusplus
 
-int flitdb_insert(flitdb **handler, unsigned short column_position, unsigned short row_position, const char *value)
+int flitdb_insert(flitdb **handler, unsigned long long int column_position, unsigned short row_position, const char *value)
 {
 	return flitdb_insert_const_char(handler, column_position, row_position, value);
 }
 
 #endif
 
-flitdb_extern int flitdb_insert_bool(flitdb **handler, unsigned short column_position, unsigned short row_position, bool value)
+flitdb_extern int flitdb_insert_bool(flitdb **handler, unsigned long long int column_position, unsigned short row_position, bool value)
 {
 	flitdb_insert_value_bool(handler, value);
 	return flitdb_insert_at(handler, column_position, row_position);
@@ -132,14 +132,14 @@ flitdb_extern int flitdb_insert_bool(flitdb **handler, unsigned short column_pos
 
 #ifdef __cplusplus
 
-int flitdb_insert(flitdb **handler, unsigned short column_position, unsigned short row_position, bool value)
+int flitdb_insert(flitdb **handler, unsigned long long int column_position, unsigned short row_position, bool value)
 {
 	return flitdb_insert_bool(handler, column_position, row_position, value);
 }
 
 #endif
 
-flitdb_extern int flitdb_delete(flitdb **handler, unsigned short column_position, unsigned short row_position)
+flitdb_extern int flitdb_delete(flitdb **handler, unsigned long long int column_position, unsigned short row_position)
 {
 	flitdb_insert_reset(handler);
 	return flitdb_insert_at(handler, column_position, row_position);
